refactor: Move code of serverpod_cli to lib/src (#769)
