+++ tools/serverpod_cli/lib/src/analytics/analytics.dart
