+++ tools/serverpod_cli/lib/src/downloads/resource_manager.dart
