+++ tools/serverpod_cli/lib/src/generator/class_analyzer.dart
