+++ tools/serverpod_cli/lib/src/generator/class_generator.dart
