+++ tools/serverpod_cli/lib/src/generator/code_analysis_collector.dart
