+++ tools/serverpod_cli/lib/src/generator/code_cleaner.dart
