+++ tools/serverpod_cli/lib/src/generator/config.dart
