+++ tools/serverpod_cli/lib/src/generator/protocol_analyzer.dart
