+++ tools/serverpod_cli/lib/src/generator/protocol_definition.dart
