+++ tools/serverpod_cli/lib/src/generator/protocol_generator_dart.dart
