+++ tools/serverpod_cli/lib/src/insights/insights.dart
