+++ tools/serverpod_cli/lib/src/internal_tools/generate_pubspecs.dart
