+++ tools/serverpod_cli/lib/src/shared/environment.dart
