+++ tools/serverpod_cli/lib/src/util/extensions.dart
