+++ tools/serverpod_cli/lib/src/util/internal_error.dart
