+++ tools/serverpod_cli/lib/src/util/string_validators.dart
