+++ tools/serverpod_cli/lib/src/util/subdirectory_extraction.dart
