+++ tools/serverpod_cli/lib/src/util/version.dart
