+++ tools/serverpod_cli/lib/src/util/windows.dart
